+++ src/pages/Main.js
@@ -13,11 +13,11 @@ import Favicon from 'src/components/Favicon'
 import FirecloudNotification from 'src/components/FirecloudNotification'
 import IdleStatusMonitor from 'src/components/IdleStatusMonitor'
 import ImportStatus from 'src/components/ImportStatus'
-import { AuthenticatedCookieSetter } from 'src/components/runtime-common'
 import SupportRequest from 'src/components/SupportRequest'
 import { getOidcConfig } from 'src/libs/auth'
 import { PageViewReporter } from 'src/libs/events'
 import { LocationProvider, PathHashInserter, Router, TitleManager } from 'src/libs/nav'
+import { AuthenticatedCookieSetter } from 'src/pages/workspaces/workspace/analysis/runtime-common'
 
 
 const Main = () => {
