+++ src/pages/workspaces/workspace/WorkspaceContainer.js
@@ -6,9 +6,7 @@ import { ButtonPrimary, Link, spinnerOverlay } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { icon } from 'src/components/icons'
 import NewWorkspaceModal from 'src/components/NewWorkspaceModal'
-import { tools } from 'src/components/notebook-utils'
 import { locationTypes } from 'src/components/region-common'
-import { analysisTabName } from 'src/components/runtime-common'
 import { TabBar } from 'src/components/tabBars'
 import TopBar from 'src/components/TopBar'
 import { updateRecentlyViewedWorkspaces } from 'src/components/workspace-utils'
@@ -20,13 +18,15 @@ import { withErrorIgnoring, withErrorReporting } from 'src/libs/error'
 import * as Nav from 'src/libs/nav'
 import { clearNotification, notify } from 'src/libs/notifications'
 import { useCancellation, useOnMount, usePrevious, useStore, withDisplayName } from 'src/libs/react-utils'
-import {
-  defaultLocation, getConvertedRuntimeStatus, getCurrentApp, getCurrentRuntime, getDiskAppType, isAzureContext, mapToPdTypes
-} from 'src/libs/runtime-utils'
 import { workspaceStore } from 'src/libs/state'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
 import { ContextBar } from 'src/pages/workspaces/workspace/analysis/ContextBar'
+import { tools } from 'src/pages/workspaces/workspace/analysis/notebook-utils'
+import { analysisTabName } from 'src/pages/workspaces/workspace/analysis/runtime-common'
+import {
+  defaultLocation, getConvertedRuntimeStatus, getCurrentApp, getCurrentRuntime, getDiskAppType, isAzureContext, mapToPdTypes
+} from 'src/pages/workspaces/workspace/analysis/runtime-utils'
 import RuntimeManager from 'src/pages/workspaces/workspace/analysis/RuntimeManager'
 import DeleteWorkspaceModal from 'src/pages/workspaces/workspace/DeleteWorkspaceModal'
 import LeaveWorkspaceModal from 'src/pages/workspaces/workspace/LeaveWorkspaceModal'
