+++ src/pages/workspaces/workspace/analysis/RuntimeManager.js
@@ -6,8 +6,6 @@ import { Fragment, useEffect, useState } from 'react'
 import { div, h, p } from 'react-hyperscript-helpers'
 import { ButtonPrimary, Clickable, Link, spinnerOverlay } from 'src/components/common'
 import Modal from 'src/components/Modal'
-import { tools } from 'src/components/notebook-utils'
-import { appLauncherTabName, GalaxyLaunchButton, GalaxyWarning } from 'src/components/runtime-common'
 import { dataSyncingDocUrl } from 'src/data/machines'
 import { Ajax } from 'src/libs/ajax'
 import { getDynamic, setDynamic } from 'src/libs/browser-storage'
@@ -16,9 +14,11 @@ import { withErrorReporting } from 'src/libs/error'
 import * as Nav from 'src/libs/nav'
 import { clearNotification, notify } from 'src/libs/notifications'
 import { useOnMount, usePrevious } from 'src/libs/react-utils'
-import { cloudProviders, getCurrentApp, getCurrentRuntime } from 'src/libs/runtime-utils'
 import { errorNotifiedApps, errorNotifiedRuntimes } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
+import { tools } from 'src/pages/workspaces/workspace/analysis/notebook-utils'
+import { appLauncherTabName, GalaxyLaunchButton, GalaxyWarning } from 'src/pages/workspaces/workspace/analysis/runtime-common'
+import { cloudProviders, getCurrentApp, getCurrentRuntime } from 'src/pages/workspaces/workspace/analysis/runtime-utils'
 
 
 export const RuntimeErrorModal = ({ runtime, onDismiss }) => {
